requirements.txt
PyAutoGUI>=0.9,<0.10
PyYAML>=6,<7
